Update transforms.conf
Updated extract_cisco_ios-general-xe and force_sourcetype_for_cisco_ios-xe.  This corrects issues with host names with numbers at the end being picked up as an event_id and also adjusted for missing event_ids.  Another adjustment is what seems to be when an interface name is logged there is a space between the platform  what looks to be an interface and the process_name.  I'm working to find out if that is for sure an interface or not.